+++ packages/components-angular/src/app/examples/index.ts
@@ -2,7 +2,10 @@ import { AccordionExampleComponent } from './accordion-example.component';
 import { AgGridExampleStorefrontComponent } from './ag-grid-example-storefront.component';
 import { AgGridExampleComponent } from './ag-grid-example.component';
 import { BannerExampleComponent } from './banner-example.component';
+import { ButtonExampleFormAttributeComponent } from './button-example-form-attribute.component';
 import { ButtonExampleFormComponent } from './button-example-form.component';
+import { ButtonPureExampleFormAttributeComponent } from './button-pure-example-form-attribute.component';
+import { ButtonPureExampleFormComponent } from './button-pure-example-form.component';
 import { CanvasExampleComponent } from './canvas-example.component';
 import { CarouselExampleDynamicSlidesComponent } from './carousel-example-dynamic-slides.component';
 import { CarouselExampleEventsComponent } from './carousel-example-events.component';
@@ -10,6 +13,7 @@ import { CarouselExampleFocusOnCenterSlideComponent } from './carousel-example-f
 import { CarouselExampleJumpToSlideComponent } from './carousel-example-jump-to-slide.component';
 import { CheckboxExampleControlledComponent } from './checkbox-example-controlled.component';
 import { CheckboxExampleComponent } from './checkbox-example.component';
+import { FlyoutExampleFormComponent } from './flyout-example-form.component';
 import { FlyoutExampleComponent } from './flyout-example.component';
 import { FlyoutMultilevelExampleActiveIdentifierComponent } from './flyout-multilevel-example-active-identifier.component';
 import { FlyoutMultilevelExampleCustomContentComponent } from './flyout-multilevel-example-custom-content.component';
@@ -51,6 +55,9 @@ export const examples = [
   AgGridExampleStorefrontComponent,
   BannerExampleComponent,
   ButtonExampleFormComponent,
+  ButtonExampleFormAttributeComponent,
+  ButtonPureExampleFormComponent,
+  ButtonPureExampleFormAttributeComponent,
   CanvasExampleComponent,
   CarouselExampleDynamicSlidesComponent,
   CarouselExampleFocusOnCenterSlideComponent,
@@ -59,6 +66,7 @@ export const examples = [
   CheckboxExampleControlledComponent,
   CheckboxExampleComponent,
   FlyoutExampleComponent,
+  FlyoutExampleFormComponent,
   FlyoutMultilevelExampleComponent,
   FlyoutMultilevelExampleActiveIdentifierComponent,
   FlyoutMultilevelExampleCustomContentComponent,
@@ -99,6 +107,9 @@ export * from './ag-grid-example.component';
 export * from './ag-grid-example-storefront.component';
 export * from './banner-example.component';
 export * from './button-example-form.component';
+export * from './button-example-form-attribute.component';
+export * from './button-pure-example-form.component';
+export * from './button-pure-example-form-attribute.component';
 export * from './canvas-example.component';
 export * from './carousel-example-dynamic-slides.component';
 export * from './carousel-example-focus-on-center-slide.component';
@@ -107,6 +118,7 @@ export * from './carousel-example-jump-to-slide.component';
 export * from './checkbox-example.component';
 export * from './checkbox-example-controlled.component';
 export * from './flyout-example.component';
+export * from './flyout-example-form.component';
 export * from './flyout-multilevel-example.component';
 export * from './flyout-multilevel-example-active-identifier.component';
 export * from './flyout-multilevel-example-custom-content.component';