+++ packages/components-js/tests/e2e/specs/button.e2e.ts
@@ -1,9 +1,10 @@
-import { expect, Locator, test, type Page } from '@playwright/test';
+import { Locator, type Page, expect, test } from '@playwright/test';
 import {
-  addEventListener,
   ClickableTests,
+  addEventListener,
   getActiveElementId,
   getEventSummary,
+  getFormDataValue,
   getLifecycleStatus,
   hasFocus,
   setContentWithDesignSystem,
@@ -17,6 +18,7 @@ skipInBrowsers(['firefox', 'webkit']);
 
 const getHost = (page: Page) => page.locator('p-button');
 const getButton = (page: Page) => page.locator('p-button button');
+const getForm = (page: Page) => page.locator('form');
 
 const initButton = (page: Page, opts?: { isLoading?: boolean; isDisabled?: boolean }): Promise<void> => {
   const { isLoading = false, isDisabled = false } = opts || {};
@@ -84,7 +86,7 @@ test('should dispatch correct click events', async ({ page }) => {
   }
 });
 
-test.describe('within form', () => {
+test.describe('form', () => {
   test("should submit parent form on click if it's type submit", async ({ page }) => {
     await setContentWithDesignSystem(
       page,
@@ -102,6 +104,64 @@ test.describe('within form', () => {
 
     await waitForImproveButtonHandlingForCustomElement(page);
     expect((await getEventSummary(form, 'submit')).counter).toBe(2);
+  });
+
+  test('Should include name and associated value in FormData on click, if the submit button is outside the form', async ({
+    page,
+  }) => {
+    const value = 'Some value';
+    const formId = 'myForm';
+    const name = 'some-name';
+
+    await setContentWithDesignSystem(
+      page,
+      `<form onsubmit="return false;" id="myForm">
+        <p-textarea name="${name}" label="Some Label" value="${value}"></p-textarea>
+      </form>
+      <p-button type="submit" form="${formId}">Submit</p-button>
+    `
+    );
+
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe('Some value');
+  });
+
+  test('Should reset value on click, if the reset button is outside the form', async ({ page }) => {
+    const value = 'Some value';
+    const formId = 'myForm';
+    const name = 'some-name';
+    await setContentWithDesignSystem(
+      page,
+      `<form onsubmit="return false;" id="${formId}">
+        <p-textarea name="${name}" label="Some Label" value="${value}"></p-textarea>
+      </form>
+      <p-button type="reset" form="${formId}">Reset</p-button>
+    `
+    );
+
+    const form = getForm(page);
+    const textarea = page.locator('textarea');
+    const newValue = 'New value';
+    await textarea.fill(newValue);
+
+    await waitForStencilLifecycle(page);
+    await expect(textarea).toHaveValue(newValue);
+
+    await addEventListener(form, 'reset');
+    expect((await getEventSummary(form, 'reset')).counter).toBe(0);
+
+    await page.locator('button[type="reset"]').click();
+    await waitForStencilLifecycle(page);
+
+    expect((await getEventSummary(form, 'reset')).counter).toBe(1);
+    await expect(textarea).toHaveValue(value);
   });
 
   test('should not submit the form if default is prevented', async ({ page }) => {
@@ -164,6 +224,56 @@ test.describe('within form', () => {
     await host.click();
 
     const urlPart = `?${name}=${value}`;
+
+    await page.waitForURL(`**/*${urlPart}`);
+    // Since the data in only available via the event submitter it is easier to test it by checking the request params
+    expect(page.url()).toContain(urlPart);
+  });
+
+  test("should submit the form when a 'submit' type button outside the form is clicked, passing the button's name and value as parameters", async ({
+    page,
+  }) => {
+    const name = 'name';
+    const value = 'Value';
+    await setContentWithDesignSystem(
+      page,
+      `<form action="/packages/components-js/public" id="myForm"></form>
+    <p-button type="submit" name="${name}" value="${value}" form="myForm">Some label</p-button>`
+    );
+
+    const host = getHost(page);
+    await host.click();
+
+    const urlPart = `?${name}=${value}`;
+
+    await page.waitForURL(`**/*${urlPart}`);
+    // Since the data in only available via the event submitter it is easier to test it by checking the request params
+    expect(page.url()).toContain(urlPart);
+  });
+
+  test("Should submit the correct FormData when the button's value is updated programmatically before submission", async ({
+    page,
+  }) => {
+    const name = 'name';
+    const value = 'Value';
+    const newValue = 'NewValue';
+    await setContentWithDesignSystem(
+      page,
+      `<form action="/packages/components-js/public">
+      <p-button type="submit" name="${name}" value="${value}">Some label</p-button>
+      </form>`
+    );
+
+    const host = getHost(page);
+    await expect(host).toHaveJSProperty('value', value);
+
+    await setProperty(host, 'value', newValue);
+    await waitForStencilLifecycle(page);
+    await expect(host).toHaveJSProperty('value', newValue);
+
+    await host.click();
+
+    const urlPart = `?${name}=${newValue}`;
 
     await page.waitForURL(`**/*${urlPart}`);
     // Since the data in only available via the event submitter it is easier to test it by checking the request params